+++ lib/headers/encoded_bytes.h
@@ -20,7 +20,7 @@ struct encoded_bytes {
     void push_back(byte x);
     void push_back(ull x);
 
-    size_t get_last();
+    byte get_last();
     bool operator[](size_t ind);
     ull const& get(size_t ind);
 
@@ -29,7 +29,7 @@ struct encoded_bytes {
 
 private:
     std::vector<ull> value;
-    size_t prev_size;
+    byte prev_size;
 };
 
 #endif //HUFFMAN_ENCODED_BYTES_H
